Push new files
